pyproject.toml
[project]
name = "torchblock"
version = "0.1.1"
description = ""
authors = [
    {name = "aesthetics-of-record", email = "aaaapple123@naver.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pandas (>=2.2.3,<3.0.0)", "pillow (>=11.1.0,<12.0.0)", "tqdm (>=4.67.1,<5.0.0)", "matplotlib (>=3.10.0,<4.0.0)", "pytest (>=8.3.4,<9.0.0)"]


[project.optional-dependencies]
cpu = [
    "torch (>=2.5.0)",
    "torchvision (>=0.20.0)",
    "torchaudio (>=2.5.0)",
]

cu118 = [
    "torch (>=2.5.1)",
    "torchvision (>=0.20.1)",
    "torchaudio (>=2.5.1)",
]

cu121 = [
    "torch (>=2.5.1)",
    "torchvision (>=0.20.1)",
    "torchaudio (>=2.5.1)",
]

cu124 = [
    "torch (>=2.5.1)",
    "torchvision (>=0.20.1)",
    "torchaudio (>=2.5.1)",
]

[tool.poetry]
package-mode = true
name = "torchblock"
description = "torchblock package"
packages = [
    { include = "torchblock" }
]

[tool.poetry.dependencies]
torch = [
    { markers = "extra == 'cpu' and extra != 'cu118' and extra != 'cu121' and extra != 'cu124'", source = "pytorch-cpu"},
    { markers = "extra == 'cu118' and extra != 'cpu' and extra != 'cu121' and extra != 'cu124'", source = "pytorch-cuda118"},
    { markers = "extra == 'cu121' and extra != 'cpu' and extra != 'cu118' and extra != 'cu124'", source = "pytorch-cuda121"},
    { markers = "extra == 'cu124' and extra != 'cpu' and extra != 'cu118' and extra != 'cu121'", source = "pytorch-cuda124"},
 ]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cuda118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cuda121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cuda124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
